refactor(server): fold hooks.ts into types.ts
`app/hooks.ts` (35 lines) held 4 hook-related type derivations consumed
in the same call sites as `CoreApp` and `AgentId`/`UserId` re-exports.
Folding them into `types.ts` collapses two type files into one.

- Move `TaskAuthorizeDispatchContext`, `DispatchAdmissionResult`,
  `TaskAuthorizeDispatchHook`, `MessageAuthorizeContext`,
  `MessageAuthorizeResult`, `MessageAuthorizeHook` into
  `app/types.ts`.
- Update 4 import sites (`app-host.ts`, `app-host.test.ts`,
  `app-host.remote.test.ts`, `layers.ts`) from `./hooks.js` →
  `./types.js`.
- Delete `app/hooks.ts`.
