+++ packages/server/src/app/app-host.test.ts
@@ -16,7 +16,7 @@ import type { Db } from "../db/client.js";
 import type { ConnectionManager } from "../transport/connection.js";
 import { makeFakeService } from "../test-utils/fakes.js";
 import { AppHost } from "./app-host.js";
-import type { MessageAuthorizeContext } from "./hooks.js";
+import type { MessageAuthorizeContext } from "./types.js";
 
 function makeAppHost(db: Db = makeEmptyDb()): { host: AppHost } {
   const connections = makeFakeService<ConnectionManager>(
